added tests for negative prices

+++ eelib/matcher.cpp
@@ -1,7 +1,7 @@
 #include "matcher.h"
 
 void Matcher::placeOrder(const Order& order){
-    if(order.price < minPrice || order.price > (minPrice + priceRange)){
+    if(order.price < minPrice || order.price > (minPrice + (int)priceRange)){
         throw new std::logic_error("order.price is outside of matcher price range");
     };
